[fixed] 主题价格不能传空问题

+++ app/Commands/Thread/CreateThread.php
@@ -97,7 +97,7 @@ class CreateThread
         // 发布长文时记录标题及价格
         if ($thread->is_long_article) {
             $thread->title = Arr::get($this->data, 'attributes.title');
-            $thread->price = Arr::get($this->data, 'attributes.price', 0);
+            $thread->price = (float) Arr::get($this->data, 'attributes.price', 0);
         }
 
         $thread->setRelation('user', $this->actor);
